move to dedicated module of helper functions
